LPD-38845 Remove feature flag, remove unused, simplify
